Poverty Probability Index (PPI) lookup table for Ghana using poverty definitions deflated with the change in 100% of national poverty line

Poverty Probability Index (PPI) lookup table for Ghana using poverty definitions deflated with the change in 100% of national poverty line

Usage

ppiGHA2015_b

Format

A data frame with 8 columns and 101 rows:

score

PPI score

ppp125

Below $1.25 per day purchasing power parity (2005)

ppp200

Below $2.00 per day purchasing power parity (2005)

ppp250

Below $2.50 per day purchasing power parity (2005)

ppp375

Below $3.75 per day purchasing power parity (2005)

ppp500

Below $5.00 per day purchasing power parity (2005)

ppp190

Below $1.90 per day purchasing power parity (2011)

ppp310

Below $3.10 per day purchasing power parity (2011)

Source

https://www.povertyindex.org

Examples

  # Access Ghana PPI table
  ppiGHA2015_b
#>     score ppp125 ppp200 ppp250 ppp375 ppp500 ppp190 ppp310
#> 0       0   83.9   96.7   97.9  100.0  100.0   89.0   99.1
#> 1       1   83.9   96.7   97.9  100.0  100.0   89.0   99.1
#> 2       2   83.9   96.7   97.9  100.0  100.0   89.0   99.1
#> 3       3   83.9   96.7   97.9  100.0  100.0   89.0   99.1
#> 4       4   83.9   96.7   97.9  100.0  100.0   89.0   99.1
#> 5       5   83.9   96.7   97.9  100.0  100.0   89.0   99.1
#> 6       6   83.9   96.7   97.9  100.0  100.0   89.0   99.1
#> 7       7   83.9   96.7   97.9  100.0  100.0   89.0   99.1
#> 8       8   83.9   96.7   97.9  100.0  100.0   89.0   99.1
#> 9       9   83.9   96.7   97.9  100.0  100.0   89.0   99.1
#> 10     10   59.6   84.5   93.5   97.9   99.4   74.9   95.2
#> 11     11   59.6   84.5   93.5   97.9   99.4   74.9   95.2
#> 12     12   59.6   84.5   93.5   97.9   99.4   74.9   95.2
#> 13     13   59.6   84.5   93.5   97.9   99.4   74.9   95.2
#> 14     14   59.6   84.5   93.5   97.9   99.4   74.9   95.2
#> 15     15   47.3   78.7   88.2   96.4   98.9   65.6   91.2
#> 16     16   47.3   78.7   88.2   96.4   98.9   65.6   91.2
#> 17     17   47.3   78.7   88.2   96.4   98.9   65.6   91.2
#> 18     18   47.3   78.7   88.2   96.4   98.9   65.6   91.2
#> 19     19   47.3   78.7   88.2   96.4   98.9   65.6   91.2
#> 20     20   39.4   75.1   86.1   96.3   98.7   62.4   87.9
#> 21     21   39.4   75.1   86.1   96.3   98.7   62.4   87.9
#> 22     22   39.4   75.1   86.1   96.3   98.7   62.4   87.9
#> 23     23   39.4   75.1   86.1   96.3   98.7   62.4   87.9
#> 24     24   39.4   75.1   86.1   96.3   98.7   62.4   87.9
#> 25     25   31.1   66.0   79.4   94.4   97.3   50.8   82.1
#> 26     26   31.1   66.0   79.4   94.4   97.3   50.8   82.1
#> 27     27   31.1   66.0   79.4   94.4   97.3   50.8   82.1
#> 28     28   31.1   66.0   79.4   94.4   97.3   50.8   82.1
#> 29     29   31.1   66.0   79.4   94.4   97.3   50.8   82.1
#> 30     30   20.3   51.1   69.4   87.5   95.4   35.4   71.6
#> 31     31   20.3   51.1   69.4   87.5   95.4   35.4   71.6
#> 32     32   20.3   51.1   69.4   87.5   95.4   35.4   71.6
#> 33     33   20.3   51.1   69.4   87.5   95.4   35.4   71.6
#> 34     34   20.3   51.1   69.4   87.5   95.4   35.4   71.6
#> 35     35   10.6   39.8   54.9   79.6   90.2   25.8   59.1
#> 36     36   10.6   39.8   54.9   79.6   90.2   25.8   59.1
#> 37     37   10.6   39.8   54.9   79.6   90.2   25.8   59.1
#> 38     38   10.6   39.8   54.9   79.6   90.2   25.8   59.1
#> 39     39   10.6   39.8   54.9   79.6   90.2   25.8   59.1
#> 40     40    7.8   27.5   42.0   74.5   85.7   17.3   46.1
#> 41     41    7.8   27.5   42.0   74.5   85.7   17.3   46.1
#> 42     42    7.8   27.5   42.0   74.5   85.7   17.3   46.1
#> 43     43    7.8   27.5   42.0   74.5   85.7   17.3   46.1
#> 44     44    7.8   27.5   42.0   74.5   85.7   17.3   46.1
#> 45     45    3.5   18.8   31.3   64.3   79.7   10.3   36.7
#> 46     46    3.5   18.8   31.3   64.3   79.7   10.3   36.7
#> 47     47    3.5   18.8   31.3   64.3   79.7   10.3   36.7
#> 48     48    3.5   18.8   31.3   64.3   79.7   10.3   36.7
#> 49     49    3.5   18.8   31.3   64.3   79.7   10.3   36.7
#> 50     50    1.5   12.6   22.5   49.4   68.2    4.7   25.8
#> 51     51    1.5   12.6   22.5   49.4   68.2    4.7   25.8
#> 52     52    1.5   12.6   22.5   49.4   68.2    4.7   25.8
#> 53     53    1.5   12.6   22.5   49.4   68.2    4.7   25.8
#> 54     54    1.5   12.6   22.5   49.4   68.2    4.7   25.8
#> 55     55    1.5    7.0   15.5   41.7   60.3    3.6   18.1
#> 56     56    1.5    7.0   15.5   41.7   60.3    3.6   18.1
#> 57     57    1.5    7.0   15.5   41.7   60.3    3.6   18.1
#> 58     58    1.5    7.0   15.5   41.7   60.3    3.6   18.1
#> 59     59    1.5    7.0   15.5   41.7   60.3    3.6   18.1
#> 60     60    0.7    4.0    9.0   31.3   52.1    1.5   11.3
#> 61     61    0.7    4.0    9.0   31.3   52.1    1.5   11.3
#> 62     62    0.7    4.0    9.0   31.3   52.1    1.5   11.3
#> 63     63    0.7    4.0    9.0   31.3   52.1    1.5   11.3
#> 64     64    0.7    4.0    9.0   31.3   52.1    1.5   11.3
#> 65     65    0.1    1.9    5.6   22.4   42.6    0.5    6.6
#> 66     66    0.1    1.9    5.6   22.4   42.6    0.5    6.6
#> 67     67    0.1    1.9    5.6   22.4   42.6    0.5    6.6
#> 68     68    0.1    1.9    5.6   22.4   42.6    0.5    6.6
#> 69     69    0.1    1.9    5.6   22.4   42.6    0.5    6.6
#> 70     70    0.1    1.7    3.7   15.5   31.9    0.4    4.3
#> 71     71    0.1    1.7    3.7   15.5   31.9    0.4    4.3
#> 72     72    0.1    1.7    3.7   15.5   31.9    0.4    4.3
#> 73     73    0.1    1.7    3.7   15.5   31.9    0.4    4.3
#> 74     74    0.1    1.7    3.7   15.5   31.9    0.4    4.3
#> 75     75    0.0    0.3    0.8    7.7   18.3    0.0    1.6
#> 76     76    0.0    0.3    0.8    7.7   18.3    0.0    1.6
#> 77     77    0.0    0.3    0.8    7.7   18.3    0.0    1.6
#> 78     78    0.0    0.3    0.8    7.7   18.3    0.0    1.6
#> 79     79    0.0    0.3    0.8    7.7   18.3    0.0    1.6
#> 80     80    0.0    0.1    0.5    2.2    6.9    0.0    0.7
#> 81     81    0.0    0.1    0.5    2.2    6.9    0.0    0.7
#> 82     82    0.0    0.1    0.5    2.2    6.9    0.0    0.7
#> 83     83    0.0    0.1    0.5    2.2    6.9    0.0    0.7
#> 84     84    0.0    0.1    0.5    2.2    6.9    0.0    0.7
#> 85     85    0.0    0.0    0.2    0.7    4.4    0.0    0.2
#> 86     86    0.0    0.0    0.2    0.7    4.4    0.0    0.2
#> 87     87    0.0    0.0    0.2    0.7    4.4    0.0    0.2
#> 88     88    0.0    0.0    0.2    0.7    4.4    0.0    0.2
#> 89     89    0.0    0.0    0.2    0.7    4.4    0.0    0.2
#> 90     90    0.0    0.0    0.0    0.2    1.2    0.0    0.0
#> 91     91    0.0    0.0    0.0    0.2    1.2    0.0    0.0
#> 92     92    0.0    0.0    0.0    0.2    1.2    0.0    0.0
#> 93     93    0.0    0.0    0.0    0.2    1.2    0.0    0.0
#> 94     94    0.0    0.0    0.0    0.2    1.2    0.0    0.0
#> 95     95    0.0    0.0    0.0    0.0    0.0    0.0    0.0
#> 96     96    0.0    0.0    0.0    0.0    0.0    0.0    0.0
#> 97     97    0.0    0.0    0.0    0.0    0.0    0.0    0.0
#> 98     98    0.0    0.0    0.0    0.0    0.0    0.0    0.0
#> 99     99    0.0    0.0    0.0    0.0    0.0    0.0    0.0
#> 100   100    0.0    0.0    0.0    0.0    0.0    0.0    0.0

  # Given a specific PPI score (from 0 - 100), get the row of poverty
  # probabilities from PPI table it corresponds to
  ppiScore <- 50
  ppiGHA2015_b[ppiGHA2015_b$score == ppiScore, ]
#>    score ppp125 ppp200 ppp250 ppp375 ppp500 ppp190 ppp310
#> 50    50    1.5   12.6   22.5   49.4   68.2    4.7   25.8

  # Use subset() function to get the row of poverty probabilities corresponding
  # to specific PPI score
  ppiScore <- 50
  subset(ppiGHA2015_b, score == ppiScore)
#>    score ppp125 ppp200 ppp250 ppp375 ppp500 ppp190 ppp310
#> 50    50    1.5   12.6   22.5   49.4   68.2    4.7   25.8

  # Given a specific PPI score (from 0 - 100), get a poverty probability
  # based on a specific poverty definition. In this example, the below $1.25
  # per day purchasing power parity (2005)
  ppiScore <- 50
  ppiGHA2015_b[ppiGHA2015_b$score == ppiScore, "ppp125"]
#> [1] 1.5