find_table

Search for PPI table by specifying region, country and/or calculation type.

find_table(region = levels(steer$region),
  country = as.character(steer$country[steer$region %in% region]),
  type = as.character(steer$type[steer$country %in% country]))

Arguments

region
Region of the world to search PPI table from. Default is c("Africa", "Asia", "Eastern Europe and Central Asia", "Latin America and the Carribean", "Middle East and North Africa"). Allows specification of one region or a vector of regions.

country
Country to search PPI table from. Default is vector of all country names from the specified region/s. Allows specification of one country name or a vector of country names.

type
Type of PPI calculation used. Can be one of two options: "sps" for the Simple Poverty Scorecard calculation or ipa for the International Poverty Alliance calculation. Default is vector of all calculation types available for the specified country/ies.

Value

A data frame in tibble format of corresponding PPI table/s matching the search parameters. The data frame contains information on the region, country, description, survey year, release year, calculation type, and filename of the returned PPI table/s.

Examples

#
# View the full data frame of all the PPI tables available through ppitables
#
find_table()

#> # A tibble: 85 x 7
#>    region    country  survey_year release_year description        type  filename
#>    <fct>     <fct>          <int>        <int> <fct>              <fct> <fct>   
#>  1 Asia      Afghani…        2007         2012 PPI for Afghanist… sps   ppiAFG2…
#>  2 Africa    Angola          2008         2015 PPI for Angola ba… sps   ppiAGO2…
#>  3 Asia      Banglad…        2010         2013 PPI for Benin bas… sps   ppiBEN2…
#>  4 Africa    Benin           2010         2012 PPI for Banglades… sps   ppiBGD2…
#>  5 Latin Am… Bolivia         2013         2015 PPI for Bolivia u… sps   ppiBOL2…
#>  6 Latin Am… Brazil          2008         2010 PPI for Brazil ba… sps   ppiBRA2…
#>  7 Africa    Burkina…        2003         2011 PPI for Burkina F… sps   ppiBFA2…
#>  8 Africa    Burkina…        2014         2017 PPI for Burkina F… sps   ppiBFA2…
#>  9 Africa    Burkina…        2014         2017 PPI for Burkina F… ipa   ppiBFA2…
#> 10 Asia      Cambodia        2011         2015 PPI for Cambodia … sps   ppiKHM2…
#> # … with 75 more rows