Poverty Probability Index (PPI) lookup table for Ghana using poverty definitions deflated with Ghana's CPI

Poverty Probability Index (PPI) lookup table for Ghana using poverty definitions deflated with Ghana's CPI

Usage

ppiGHA2015_a

Format

A data frame with 13 columns and 101 rows:

score

PPI score

nlFood

Food poverty line

nl100

National poverty line (100%)

nl150

National poverty line (150%)

nl200

National poverty line (200%)

half100

Poorest half below 100% national

ppp125

Below $1.25 per day purchasing power parity (2005)

ppp200

Below $2.00 per day purchasing power parity (2005)

ppp250

Below $2.50 per day purchasing power parity (2005)

ppp375

Below $3.75 per day purchasing power parity (2005)

ppp500

Below $5.00 per day purchasing power parity (2005)

ppp190

Below $1.90 per day purchasing power parity (2011)

ppp310

Below $3.10 per day purchasing power parity (2011)

Source

https://www.povertyindex.org

Examples

  # Access Ghana PPI table
  ppiGHA2015_a
#>     score nlFood nl100 nl150 nl200 half100 ppp125 ppp200 ppp250 ppp375 ppp500
#> 0       0   70.1  91.4  99.1 100.0    74.5   54.1   85.9   91.5   97.9  100.0
#> 1       1   70.1  91.4  99.1 100.0    74.5   54.1   85.9   91.5   97.9  100.0
#> 2       2   70.1  91.4  99.1 100.0    74.5   54.1   85.9   91.5   97.9  100.0
#> 3       3   70.1  91.4  99.1 100.0    74.5   54.1   85.9   91.5   97.9  100.0
#> 4       4   70.1  91.4  99.1 100.0    74.5   54.1   85.9   91.5   97.9  100.0
#> 5       5   70.1  91.4  99.1 100.0    74.5   54.1   85.9   91.5   97.9  100.0
#> 6       6   70.1  91.4  99.1 100.0    74.5   54.1   85.9   91.5   97.9  100.0
#> 7       7   70.1  91.4  99.1 100.0    74.5   54.1   85.9   91.5   97.9  100.0
#> 8       8   70.1  91.4  99.1 100.0    74.5   54.1   85.9   91.5   97.9  100.0
#> 9       9   70.1  91.4  99.1 100.0    74.5   54.1   85.9   91.5   97.9  100.0
#> 10     10   46.1  75.9  89.3  97.7    53.3   38.2   66.0   78.2   95.1   97.7
#> 11     11   46.1  75.9  89.3  97.7    53.3   38.2   66.0   78.2   95.1   97.7
#> 12     12   46.1  75.9  89.3  97.7    53.3   38.2   66.0   78.2   95.1   97.7
#> 13     13   46.1  75.9  89.3  97.7    53.3   38.2   66.0   78.2   95.1   97.7
#> 14     14   46.1  75.9  89.3  97.7    53.3   38.2   66.0   78.2   95.1   97.7
#> 15     15   34.6  66.8  86.9  96.4    41.3   24.5   55.2   71.5   91.2   95.9
#> 16     16   34.6  66.8  86.9  96.4    41.3   24.5   55.2   71.5   91.2   95.9
#> 17     17   34.6  66.8  86.9  96.4    41.3   24.5   55.2   71.5   91.2   95.9
#> 18     18   34.6  66.8  86.9  96.4    41.3   24.5   55.2   71.5   91.2   95.9
#> 19     19   34.6  66.8  86.9  96.4    41.3   24.5   55.2   71.5   91.2   95.9
#> 20     20   26.2  63.8  86.0  93.7    32.9   19.5   49.1   67.0   86.9   93.8
#> 21     21   26.2  63.8  86.0  93.7    32.9   19.5   49.1   67.0   86.9   93.8
#> 22     22   26.2  63.8  86.0  93.7    32.9   19.5   49.1   67.0   86.9   93.8
#> 23     23   26.2  63.8  86.0  93.7    32.9   19.5   49.1   67.0   86.9   93.8
#> 24     24   26.2  63.8  86.0  93.7    32.9   19.5   49.1   67.0   86.9   93.8
#> 25     25   18.9  53.3  80.1  92.4    28.4   12.1   38.5   57.1   81.7   92.7
#> 26     26   18.9  53.3  80.1  92.4    28.4   12.1   38.5   57.1   81.7   92.7
#> 27     27   18.9  53.3  80.1  92.4    28.4   12.1   38.5   57.1   81.7   92.7
#> 28     28   18.9  53.3  80.1  92.4    28.4   12.1   38.5   57.1   81.7   92.7
#> 29     29   18.9  53.3  80.1  92.4    28.4   12.1   38.5   57.1   81.7   92.7
#> 30     30   13.1  40.2  71.1  84.0    21.3    7.2   27.1   41.3   71.4   84.3
#> 31     31   13.1  40.2  71.1  84.0    21.3    7.2   27.1   41.3   71.4   84.3
#> 32     32   13.1  40.2  71.1  84.0    21.3    7.2   27.1   41.3   71.4   84.3
#> 33     33   13.1  40.2  71.1  84.0    21.3    7.2   27.1   41.3   71.4   84.3
#> 34     34   13.1  40.2  71.1  84.0    21.3    7.2   27.1   41.3   71.4   84.3
#> 35     35    6.9  29.0  58.1  76.4    10.7    3.7   16.2   31.4   58.8   77.4
#> 36     36    6.9  29.0  58.1  76.4    10.7    3.7   16.2   31.4   58.8   77.4
#> 37     37    6.9  29.0  58.1  76.4    10.7    3.7   16.2   31.4   58.8   77.4
#> 38     38    6.9  29.0  58.1  76.4    10.7    3.7   16.2   31.4   58.8   77.4
#> 39     39    6.9  29.0  58.1  76.4    10.7    3.7   16.2   31.4   58.8   77.4
#> 40     40    4.4  19.6  45.3  69.2     7.8    2.5   11.6   20.7   45.9   71.5
#> 41     41    4.4  19.6  45.3  69.2     7.8    2.5   11.6   20.7   45.9   71.5
#> 42     42    4.4  19.6  45.3  69.2     7.8    2.5   11.6   20.7   45.9   71.5
#> 43     43    4.4  19.6  45.3  69.2     7.8    2.5   11.6   20.7   45.9   71.5
#> 44     44    4.4  19.6  45.3  69.2     7.8    2.5   11.6   20.7   45.9   71.5
#> 45     45    1.4  11.7  35.2  58.8     4.0    0.6    5.3   12.8   36.4   59.6
#> 46     46    1.4  11.7  35.2  58.8     4.0    0.6    5.3   12.8   36.4   59.6
#> 47     47    1.4  11.7  35.2  58.8     4.0    0.6    5.3   12.8   36.4   59.6
#> 48     48    1.4  11.7  35.2  58.8     4.0    0.6    5.3   12.8   36.4   59.6
#> 49     49    1.4  11.7  35.2  58.8     4.0    0.6    5.3   12.8   36.4   59.6
#> 50     50    0.9   7.2  25.6  47.3     2.0    0.4    2.5    6.8   25.2   45.1
#> 51     51    0.9   7.2  25.6  47.3     2.0    0.4    2.5    6.8   25.2   45.1
#> 52     52    0.9   7.2  25.6  47.3     2.0    0.4    2.5    6.8   25.2   45.1
#> 53     53    0.9   7.2  25.6  47.3     2.0    0.4    2.5    6.8   25.2   45.1
#> 54     54    0.9   7.2  25.6  47.3     2.0    0.4    2.5    6.8   25.2   45.1
#> 55     55    0.9   4.3  17.7  35.7     1.7    0.3    2.2    5.1   17.6   36.7
#> 56     56    0.9   4.3  17.7  35.7     1.7    0.3    2.2    5.1   17.6   36.7
#> 57     57    0.9   4.3  17.7  35.7     1.7    0.3    2.2    5.1   17.6   36.7
#> 58     58    0.9   4.3  17.7  35.7     1.7    0.3    2.2    5.1   17.6   36.7
#> 59     59    0.9   4.3  17.7  35.7     1.7    0.3    2.2    5.1   17.6   36.7
#> 60     60    0.2   2.2  10.6  26.5     0.7    0.0    1.0    2.1   10.9   25.9
#> 61     61    0.2   2.2  10.6  26.5     0.7    0.0    1.0    2.1   10.9   25.9
#> 62     62    0.2   2.2  10.6  26.5     0.7    0.0    1.0    2.1   10.9   25.9
#> 63     63    0.2   2.2  10.6  26.5     0.7    0.0    1.0    2.1   10.9   25.9
#> 64     64    0.2   2.2  10.6  26.5     0.7    0.0    1.0    2.1   10.9   25.9
#> 65     65    0.0   1.1   7.4  21.2     0.2    0.0    0.2    0.7    6.5   17.9
#> 66     66    0.0   1.1   7.4  21.2     0.2    0.0    0.2    0.7    6.5   17.9
#> 67     67    0.0   1.1   7.4  21.2     0.2    0.0    0.2    0.7    6.5   17.9
#> 68     68    0.0   1.1   7.4  21.2     0.2    0.0    0.2    0.7    6.5   17.9
#> 69     69    0.0   1.1   7.4  21.2     0.2    0.0    0.2    0.7    6.5   17.9
#> 70     70    0.0   0.8   4.6  13.3     0.1    0.0    0.1    0.6    4.3   11.7
#> 71     71    0.0   0.8   4.6  13.3     0.1    0.0    0.1    0.6    4.3   11.7
#> 72     72    0.0   0.8   4.6  13.3     0.1    0.0    0.1    0.6    4.3   11.7
#> 73     73    0.0   0.8   4.6  13.3     0.1    0.0    0.1    0.6    4.3   11.7
#> 74     74    0.0   0.8   4.6  13.3     0.1    0.0    0.1    0.6    4.3   11.7
#> 75     75    0.0   0.3   1.4   6.5     0.0    0.0    0.0    0.0    1.6    6.2
#> 76     76    0.0   0.3   1.4   6.5     0.0    0.0    0.0    0.0    1.6    6.2
#> 77     77    0.0   0.3   1.4   6.5     0.0    0.0    0.0    0.0    1.6    6.2
#> 78     78    0.0   0.3   1.4   6.5     0.0    0.0    0.0    0.0    1.6    6.2
#> 79     79    0.0   0.3   1.4   6.5     0.0    0.0    0.0    0.0    1.6    6.2
#> 80     80    0.0   0.0   0.6   1.1     0.0    0.0    0.0    0.0    0.7    1.3
#> 81     81    0.0   0.0   0.6   1.1     0.0    0.0    0.0    0.0    0.7    1.3
#> 82     82    0.0   0.0   0.6   1.1     0.0    0.0    0.0    0.0    0.7    1.3
#> 83     83    0.0   0.0   0.6   1.1     0.0    0.0    0.0    0.0    0.7    1.3
#> 84     84    0.0   0.0   0.6   1.1     0.0    0.0    0.0    0.0    0.7    1.3
#> 85     85    0.0   0.0   0.2   0.4     0.0    0.0    0.0    0.0    0.2    0.4
#> 86     86    0.0   0.0   0.2   0.4     0.0    0.0    0.0    0.0    0.2    0.4
#> 87     87    0.0   0.0   0.2   0.4     0.0    0.0    0.0    0.0    0.2    0.4
#> 88     88    0.0   0.0   0.2   0.4     0.0    0.0    0.0    0.0    0.2    0.4
#> 89     89    0.0   0.0   0.2   0.4     0.0    0.0    0.0    0.0    0.2    0.4
#> 90     90    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#> 91     91    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#> 92     92    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#> 93     93    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#> 94     94    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#> 95     95    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#> 96     96    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#> 97     97    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#> 98     98    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#> 99     99    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#> 100   100    0.0   0.0   0.0   0.0     0.0    0.0    0.0    0.0    0.0    0.0
#>     ppp190 ppp310
#> 0     74.5   96.7
#> 1     74.5   96.7
#> 2     74.5   96.7
#> 3     74.5   96.7
#> 4     74.5   96.7
#> 5     74.5   96.7
#> 6     74.5   96.7
#> 7     74.5   96.7
#> 8     74.5   96.7
#> 9     74.5   96.7
#> 10    50.7   81.7
#> 11    50.7   81.7
#> 12    50.7   81.7
#> 13    50.7   81.7
#> 14    50.7   81.7
#> 15    39.9   74.7
#> 16    39.9   74.7
#> 17    39.9   74.7
#> 18    39.9   74.7
#> 19    39.9   74.7
#> 20    32.6   70.6
#> 21    32.6   70.6
#> 22    32.6   70.6
#> 23    32.6   70.6
#> 24    32.6   70.6
#> 25    25.1   62.1
#> 26    25.1   62.1
#> 27    25.1   62.1
#> 28    25.1   62.1
#> 29    25.1   62.1
#> 30    16.5   45.4
#> 31    16.5   45.4
#> 32    16.5   45.4
#> 33    16.5   45.4
#> 34    16.5   45.4
#> 35     7.9   35.4
#> 36     7.9   35.4
#> 37     7.9   35.4
#> 38     7.9   35.4
#> 39     7.9   35.4
#> 40     5.1   24.0
#> 41     5.1   24.0
#> 42     5.1   24.0
#> 43     5.1   24.0
#> 44     5.1   24.0
#> 45     2.4   16.1
#> 46     2.4   16.1
#> 47     2.4   16.1
#> 48     2.4   16.1
#> 49     2.4   16.1
#> 50     1.0    9.5
#> 51     1.0    9.5
#> 52     1.0    9.5
#> 53     1.0    9.5
#> 54     1.0    9.5
#> 55     0.9    5.8
#> 56     0.9    5.8
#> 57     0.9    5.8
#> 58     0.9    5.8
#> 59     0.9    5.8
#> 60     0.4    2.9
#> 61     0.4    2.9
#> 62     0.4    2.9
#> 63     0.4    2.9
#> 64     0.4    2.9
#> 65     0.1    1.3
#> 66     0.1    1.3
#> 67     0.1    1.3
#> 68     0.1    1.3
#> 69     0.1    1.3
#> 70     0.1    0.8
#> 71     0.1    0.8
#> 72     0.1    0.8
#> 73     0.1    0.8
#> 74     0.1    0.8
#> 75     0.0    0.1
#> 76     0.0    0.1
#> 77     0.0    0.1
#> 78     0.0    0.1
#> 79     0.0    0.1
#> 80     0.0    0.0
#> 81     0.0    0.0
#> 82     0.0    0.0
#> 83     0.0    0.0
#> 84     0.0    0.0
#> 85     0.0    0.0
#> 86     0.0    0.0
#> 87     0.0    0.0
#> 88     0.0    0.0
#> 89     0.0    0.0
#> 90     0.0    0.0
#> 91     0.0    0.0
#> 92     0.0    0.0
#> 93     0.0    0.0
#> 94     0.0    0.0
#> 95     0.0    0.0
#> 96     0.0    0.0
#> 97     0.0    0.0
#> 98     0.0    0.0
#> 99     0.0    0.0
#> 100    0.0    0.0

  # Given a specific PPI score (from 0 - 100), get the row of poverty
  # probabilities from PPI table it corresponds to
  ppiScore <- 50
  ppiGHA2015_a[ppiGHA2015_a$score == ppiScore, ]
#>    score nlFood nl100 nl150 nl200 half100 ppp125 ppp200 ppp250 ppp375 ppp500
#> 50    50    0.9   7.2  25.6  47.3       2    0.4    2.5    6.8   25.2   45.1
#>    ppp190 ppp310
#> 50      1    9.5

  # Use subset() function to get the row of poverty probabilities corresponding
  # to specific PPI score
  ppiScore <- 50
  subset(ppiGHA2015_a, score == ppiScore)
#>    score nlFood nl100 nl150 nl200 half100 ppp125 ppp200 ppp250 ppp375 ppp500
#> 50    50    0.9   7.2  25.6  47.3       2    0.4    2.5    6.8   25.2   45.1
#>    ppp190 ppp310
#> 50      1    9.5

  # Given a specific PPI score (from 0 - 100), get a poverty probability
  # based on a specific poverty definition. In this example, the national
  # poverty line definition
  ppiScore <- 50
  ppiGHA2015_a[ppiGHA2015_a$score == ppiScore, "nl100"]
#> [1] 7.2